refactor: remove `json-docs` and renamed it to `docTypes.ts`
