+++ packages/luals-annotations/src/luaLSType.ts
@@ -1,4 +1,4 @@
-import { Docs } from "./json-docs";
+import { Docs } from "./docTypes";
 
 function convertFunctionProperty(property: Docs.FunctionProperty): string {
 	const params = property.parameters ?? [];
